Deduplicate naming functions
