+++ tests/names_test.py
@@ -3,7 +3,7 @@ import unittest
 from cryptography import x509
 from cryptography.x509.oid import NameOID
 
-from lib.dn import as_name
+from lib.names import as_name
 
 
 class TestDNFunctions(unittest.TestCase):
@@ -14,7 +14,7 @@ class TestDNFunctions(unittest.TestCase):
             'C': 'NL',
             'O': 'Organization',
             'organizationIdentifier': '12345',
-            'CN': 'Test User',
+            '2.5.4.3': 'Test User',
         }
 
         actual = as_name(dn_input)
@@ -27,6 +27,15 @@ class TestDNFunctions(unittest.TestCase):
 
         self.assertEqual(actual, expected)
         self.assertEqual(actual.rfc4514_string(), '2.5.4.97=12345,CN=Test User,O=Organization,C=NL')
+
+    def test_incorrect_name(self):
+        dn_input = {
+            'C': 'NL',
+            'unknown': 'Organization',
+        }
+
+        with self.assertRaises(ValueError):
+            as_name(dn_input)
 
 
 if __name__ == '__main__':
